# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UKM Politeknik",
        "description": "API UKM Politeknik Negeri Lampung. Response dapat berubah sewaktu waktu menyesuaikan perbaikan",
        "contact": {
            "email": "leopradana013@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/private/agenda": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Daftar Agenda",
                "operationId": "61d603644b521cbfde0c770b809416e3",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Membuat Agenda",
                "description": "Membuat informasi agenda",
                "operationId": "c955bd9303bd8c5df21eaebda5013caa",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "content",
                                    "lokasi",
                                    "tanggal",
                                    "waktu"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Gambar agenda",
                                        "type": "file"
                                    },
                                    "title": {
                                        "description": "Judul informasi agenda",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Isi agenda (Use WYSWIG)",
                                        "type": "string"
                                    },
                                    "lokasi": {
                                        "description": "Lokasi agenda",
                                        "type": "string"
                                    },
                                    "tanggal": {
                                        "description": "Tanggal dimulai. Ex: YYYY-MM-DD (SQL Format Default)",
                                        "type": "string"
                                    },
                                    "waktu": {
                                        "description": "Waktu dimulai. Ex: h:mm (Moment JS Format)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/private/agenda/{agenda_id}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Detail agenda",
                "description": "Informasi lengkap data agenda",
                "operationId": "e4a0dbef80bcdbdd290fe2f059113626",
                "parameters": [
                    {
                        "name": "agenda_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Perbarui agenda",
                "description": "Perbarui informasi agenda",
                "operationId": "41e1d27ff1ddcb0fa5e0ba42abfa72ce",
                "parameters": [
                    {
                        "name": "agenda_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "content",
                                    "lokasi",
                                    "tanggal",
                                    "waktu"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Gambar agenda",
                                        "type": "file"
                                    },
                                    "title": {
                                        "description": "Judul informasi agenda",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Isi agenda (Use WYSWIG)",
                                        "type": "string"
                                    },
                                    "lokasi": {
                                        "description": "Lokasi agenda",
                                        "type": "string"
                                    },
                                    "tanggal": {
                                        "description": "Tanggal dimulai. Ex: YYYY-MM-DD (SQL Format Default)",
                                        "type": "string"
                                    },
                                    "waktu": {
                                        "description": "Waktu dimulai. Ex: h:mm (Moment JS Format)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Menghapus agenda",
                "description": "Menghapus informasi agenda",
                "operationId": "3b47737c3684d0ec53854a8ad16d0174",
                "parameters": [
                    {
                        "name": "agenda_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}